requirements.txt
Django==2.1.1
django-bootstrap-datepicker==1.3
django-crispy-forms==1.7.2
djangorestframework==3.8.2
gunicorn==19.9.0
Pillow==5.2.0
pytz==2018.5
